+++ src/main/java/janggi/dao/BoardVO.java
@@ -1,9 +1,10 @@
 package janggi.dao;
 
-import janggi.board.Board;
-import janggi.piece.Camp;
-import janggi.piece.Piece;
-import janggi.position.Position;
+import janggi.domain.board.Board;
+import janggi.exception.DataAccessException;
+import janggi.domain.piece.Camp;
+import janggi.domain.piece.Piece;
+import janggi.domain.position.Position;
 import java.util.Map;
 
 public record BoardVO(Long id, String teamCode, String currentCamp, String winnerCamp) {
@@ -18,6 +19,9 @@ public record BoardVO(Long id, String teamCode, String currentCamp, String winne
     }
 
     public Board toBoard(Map<Position, Piece> cells) {
+        if (cells == null) {
+            throw new DataAccessException("보드판은 null이 될 수 없습니다.");
+        }
         return new Board(
                 cells,
                 Camp.valueOf(currentCamp)
