+++ src/main/java/janggi/domain/piece/Chariot.java
@@ -1,9 +1,9 @@
-package janggi.piece;
+package janggi.domain.piece;
 
-import janggi.board.Board;
+import janggi.domain.board.Board;
 import janggi.exception.ErrorException;
-import janggi.position.Movement;
-import janggi.position.Position;
+import janggi.domain.position.Movement;
+import janggi.domain.position.Position;
 import java.util.Set;
 
 public final class Chariot extends Piece {
