+++ src/main/java/janggi/domain/piece/Elephant.java
@@ -1,9 +1,9 @@
-package janggi.piece;
+package janggi.domain.piece;
 
-import janggi.board.Board;
+import janggi.domain.board.Board;
 import janggi.exception.ErrorException;
-import janggi.position.Movement;
-import janggi.position.Position;
+import janggi.domain.position.Movement;
+import janggi.domain.position.Position;
 import java.util.HashSet;
 import java.util.Set;
