+++ src/main/java/janggi/domain/piece/Empty.java
@@ -1,7 +1,7 @@
-package janggi.piece;
+package janggi.domain.piece;
 
-import janggi.board.Board;
-import janggi.position.Movement;
+import janggi.domain.board.Board;
+import janggi.domain.position.Movement;
 
 public class Empty extends Piece {
