+++ src/main/java/janggi/domain/piece/Point.java
@@ -1,4 +1,4 @@
-package janggi.piece;
+package janggi.domain.piece;
 
 public enum Point {
