+++ src/test/java/janggi/position/MovementTest.java
@@ -3,6 +3,8 @@ package janggi.position;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import janggi.domain.position.Movement;
+import janggi.domain.position.Position;
 import janggi.exception.ErrorException;
 import java.util.Set;
 import org.junit.jupiter.api.DisplayName;
